Fix some more  (MSCV) compiler warnings
reg. `size_t` vs. `int`, fixes 271 warnings

+++ src/engraving/libmscore/instrument.cpp
@@ -1498,7 +1498,7 @@ void Instrument::addShortName(const StaffName& f)
     _shortNames.push_back(f);
 }
 
-int Instrument::cleffTypeCount() const
+size_t Instrument::cleffTypeCount() const
 {
     return _clefType.size();
 }
